pyproject.toml
[project]
name = "research-agent-fixed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.56",
    "ipykernel>=7.2.0",
    "langchain-aws>=1.3.0",
    "langchain-core>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.9",
    "langgraph-cli[inmem]>=0.4.12",
    "pydantic>=2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
